Bug fixing
MIP detection efficiency is defined with a nominator of the number of tracks with 8 efficient chambers, but currently this is common to all tested chambers. However for a specific tested chamber, only it can have more than two hits. For example: currently chamber 5 can have 5 hits when we test chamber 7.

Fixing isMIP and efficiency estimation